♻️refactor: 외부 연동 관련 infra 패키지 분리
외부 서버 연동 관리를 위한 /global/infra 패키지 구축 및 카카오 인증 infra 경로 이동
